+++ clang-tools-extra/clang-tidy/cheri/CheriTidyModule.cpp
@@ -9,6 +9,7 @@
 #include "../ClangTidy.h"
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
+#include "FixdriverdataCheck.h"
 #include "FixinttoptrcastCheck.h"
 #include "FixptrtoulongcastCheck.h"
 #include "IoctlCheck.h"
@@ -21,6 +22,8 @@ namespace cheri {
 class CheriModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
+    CheckFactories.registerCheck<FixdriverdataCheck>(
+        "cheri-FixDriverData");
     CheckFactories.registerCheck<FixinttoptrcastCheck>(
         "cheri-FixIntToPtrCast");
     CheckFactories.registerCheck<FixptrtoulongcastCheck>(